// src/app/models/emo.ts
import { Personal } from "./personal";
import { Proemo } from "./proemo";

export class Emo {
    id: number;
    codigo: string;
    codempresa: string;
    sede: string;
    codprogram: number;
    codempresapers: string;
    codpersonal: string;
    codtipoeva: string;
    codproveedor: string;
    codempresacarg: string;
    codcategoria: string;
    codcargo: string;
    fecha: string;
    fechaprogram: string;
    fechavencimiento: string;
    codestado: boolean;

    antpersonales: string;
    antfamiliares: string;
    alergiasram: string;
    codgrpfact: number;
    codpressis: number;
    codpresdia: number;
    peso: number;
    talla: number;
    imc: number;
    codestnutri: number;
    dxmuscesq: string;
    dxraditor: string;
    indineum: string;
    audioider: string;
    audioiizq: string;
    dxexpiro: string;
    hemoglo: string;
    hemogra: string;
    glucosa: string;
    tgo: string;
    tgp: string;
    colesterol: string;
    hdl: string;
    ldl: string;
    vldl: string;
    triglice: string;
    riescoron: string;
    examcomorina: string;
    dxoftalgene: string;
    agudviscerca: string;
    agudvislejos: string;
    refraccion: string;
    estereop: string;
    testish: string;
    dxelectrocar: string;
    dxpsicologia: string;
    dxexammedgen: string;
    resummed: string;
    resummedclinica: string;
    recomend: string;
    recomendclinica: string;
    codaptitud: number;
    restricciones: string;
    fechavigencia: string;

    personal: Personal;
    personales: Personal[];
    programa: Proemo;
}